+++ models/2ph_geothermal_mass_flux/model.py
@@ -1,0 +1,170 @@
+from darts.reservoirs.struct_reservoir import StructReservoir
+from darts.models.cicd_model import CICDModel
+from darts.engines import value_vector
+import numpy as np
+
+from darts.physics.super.physics import Compositional
+from darts.physics.super.property_container import PropertyContainer
+
+from darts.physics.properties.basic import ConstFunc, PhaseRelPerm
+from darts.physics.properties.density import DensityBasic
+from darts.physics.properties.enthalpy import EnthalpyBasic
+
+
+class Model(CICDModel):
+    def __init__(self, mode='rhs', well_rate=1, outflow=1000):
+        # call base class constructor
+        super().__init__()
+
+        # measure time spend on reading/initialization
+        self.timer.node["initialization"].start()
+        self.mode = mode
+        self.well_rate = well_rate
+        self.set_reservoir(mode)
+        self.set_physics()
+
+        self.set_sim_params(first_ts=0.0001, mult_ts=2, max_ts=5, runtime=1, tol_newton=1e-3, tol_linear=1e-6)
+
+        if self.mode == 'rhs':
+            # add outflux to the middle cell
+            self.set_rhs_flux(inflow_cells=np.array([self.reservoir.nx // 2]), inflow_var_idx=0, outflow=outflow)
+
+        self.timer.node["initialization"].stop()
+
+        self.initial_values = {self.physics.vars[0]: 200.,
+                               self.physics.vars[1]: 350.
+                               }
+
+    def set_reservoir(self, mode: str):
+        """Reservoir construction"""
+        # reservoir geometry： for realistic case, one just needs to load the data and input it
+        nx = 100
+        reservoir = StructReservoir(self.timer, nx=nx, ny=1, nz=1, dx=10.0, dy=10.0, dz=1, permx=5, permy=5, permz=5,
+                                    poro=0.2, hcap=2200, rcond=181.44, depth=100)
+        if mode == 'wells':
+            reservoir.add_well("P1", perf_list=[(nx//2, 1, 1)], multi_segment=False)
+
+        return super().set_reservoir(reservoir)
+
+    def set_physics(self):
+        """Physical properties"""
+        zero = 1e-13
+        components = ['w']
+        phases = ['wat', 'gas']
+
+        self.inj = value_vector([300])
+
+        property_container = ModelProperties(phases_name=phases, components_name=components, min_z=zero/10)
+
+        # Define property evaluators based on custom properties
+        property_container.density_ev = dict([('wat', DensityBasic(compr=1e-5, dens0=1014)),
+                                              ('gas', DensityBasic(compr=5e-3, dens0=50))])
+        property_container.viscosity_ev = dict([('wat', ConstFunc(0.3)),
+                                                ('gas', ConstFunc(0.03))])
+        property_container.rel_perm_ev = dict([('wat', PhaseRelPerm("oil", 0.1, 0.1)),
+                                               ('gas', PhaseRelPerm("gas", 0.1, 0.1))])
+        property_container.enthalpy_ev = dict([('wat', EnthalpyBasic(hcap=4.18)),
+                                               ('gas', EnthalpyBasic(hcap=0.035))])
+        property_container.conductivity_ev = dict([('wat', ConstFunc(1.)),
+                                                   ('gas', ConstFunc(1.))])
+        property_container.rock_energy_ev = EnthalpyBasic(hcap=1.0)
+
+        # create physics
+        thermal = True
+        physics = Compositional(components, phases, self.timer,
+                                n_points=400, min_p=0, max_p=1000, min_z=zero, max_z=1-zero,
+                                min_t=273.15 + 20, max_t=273.15 + 200, thermal=thermal)
+        physics.add_property_region(property_container)
+
+        return super().set_physics(physics)
+
+    def set_well_controls(self):
+        for i, w in enumerate(self.reservoir.wells):
+            if 'I' in w.name:
+                #w.control = self.physics.new_rate_inj(200, self.inj, 1)
+                #w.control = self.physics.new_bhp_inj(210, self.inj)
+                w.control = self.physics.new_rate_inj(self.well_rate, self.inj, 0)
+                #w.control = self.physics.new_bhp_inj(450, self.inj)
+            else:
+                w.control = self.physics.new_rate_prod(self.well_rate, iph=0)
+
+    def set_rhs_flux(self, inflow_cells: np.array, inflow_var_idx: int, outflow: float):
+        '''
+        function to specify the inflow or outflow to the cells
+        it sets up self.rhs_flux vector on nvar * ncells size
+        which will be added to rhs in darts_model.run_python function
+        :param inflow_cells: cell indices where to apply inflow or outflow
+        :param inflow_var_idx: variable index [0..nvars-1]
+        :param outflow: inflow_var_idx<nc => kMol/day, else kJ/day (thermal var)
+        if outflow < 0 then it is actually inflow
+        '''
+        nv = self.physics.n_vars
+        nb = self.mesh.n_res_blocks
+        self.rhs_flux = np.zeros(nb * nv)
+        # extract pointer to values corresponding to var_idx
+        rhs_flux_var = self.rhs_flux[inflow_var_idx::nv]
+        # set values for the cells defined in inflow_cells
+        rhs_flux_var[inflow_cells] = outflow
+
+    # overload base darts run function, because run does not apply_rhs_flux.
+    # in CI/CD only run() is called
+    def run(self):
+        self.run_python(self.runtime)
+
+
+class ModelProperties(PropertyContainer):
+    def __init__(self, phases_name, components_name, min_z=1e-11):
+        # Call base class constructor
+        self.nph = len(phases_name)
+        Mw = np.ones(self.nph)
+        super().__init__(phases_name, components_name, Mw, min_z, temperature=None)
+        self.x = np.ones((self.nph, self.nc))
+
+    def evaluate(self, state):
+        """
+        Class methods which evaluates the state operators for the element based physics
+        :param state: state variables [pres, comp_0, ..., comp_N-1]
+        :param values: values of the operators (used for storing the operator values)
+        :return: updated value for operators, stored in values
+        """
+        # Composition vector and pressure from state:
+        vec_state_as_np = np.asarray(state)
+        pressure = vec_state_as_np[0]
+
+        self.ph = [0]
+
+        for j in self.ph:
+            M = 0
+            # molar weight of mixture
+            for i in range(self.nc):
+                M += self.Mw[i] * self.x[j][i]
+            self.dens[j] = self.density_ev[self.phases_name[j]].evaluate(pressure, 0)  # output in [kg/m3]
+            self.dens_m[j] = self.dens[j] / M
+            self.mu[j] = self.viscosity_ev[self.phases_name[j]].evaluate()  # output in [cp]
+
+        self.nu[0] = 1
+        self.compute_saturation(self.ph)
+
+        for j in self.ph:
+            self.kr[j] = self.rel_perm_ev[self.phases_name[j]].evaluate(self.sat[j])
+            self.pc[j] = 0
+
+        mass_source = np.zeros(1)
+
+        return self.ph, self.sat, self.x, self.dens, self.dens_m, self.mu, self.kr, self.pc, mass_source
+
+    def evaluate_at_cond(self, pressure, zc):
+
+        self.sat[:] = 0
+
+        ph = [0]
+        for j in ph:
+            self.dens_m[j] = self.density_ev[self.phases_name[j]].evaluate(1, 0)
+
+        self.dens_m = [1025, 0.77]  # to match DO based on PVT
+
+        self.nu[0] = 1
+        self.compute_saturation(ph)
+
+        return self.sat, self.dens_m
+