+++ models/CO2_foam_CCS/model.py
@@ -1,14 +1,16 @@
-from reservoir import UnstructReservoir
+from darts.engines import *
 from darts.models.cicd_model import CICDModel
+
+from darts.reservoirs.unstruct_reservoir import UnstructReservoir
 from darts.physics.super.physics import Compositional
+from property_container import PropertyContainer
+from operator_evaluator import AccFluxGravityEvaluator, AccFluxGravityWellEvaluator, RateEvaluator, PropertyEvaluator
+
 from darts.physics.properties.basic import ConstFunc
 from darts.physics.properties.density import DensityBasic, DensityBrineCO2
 from darts.physics.properties.flash import ConstantK
 
-from property_container import PropertyContainer
 import numpy as np
-from operator_evaluator import AccFluxGravityEvaluator, AccFluxGravityWellEvaluator, RateEvaluator, PropertyEvaluator
-from darts.engines import *
 
 
 class Model(CICDModel):
@@ -37,8 +39,17 @@ class Model(CICDModel):
         const_perm = 100
         poro = 0.15
         mesh_file = 'wedgesmall.msh'
-        reservoir = UnstructReservoir(permx=const_perm, permy=const_perm, permz=const_perm,
-                                      frac_aper=0, mesh_file=mesh_file, poro=poro)
+        reservoir = UnstructReservoir(self.timer, permx=const_perm, permy=const_perm, permz=const_perm,
+                                      frac_aper=0, mesh_file=mesh_file, poro=poro, cache=False)
+
+        # Add injection well for CO2:
+        reservoir.add_well("I1", depth=5, wellbore_diameter=0.1)
+        # Perforate all boundary cells:
+        for nth_perf in range(len(self.left_boundary_cells)):
+            well_index = mesh.volume[self.left_boundary_cells[nth_perf]] / self.max_well_vol * self.well_index
+            well_indexD = 0.
+            self.add_perforation(well=self.wells[-1], res_block=self.left_boundary_cells[nth_perf],
+                                 well_index=well_index, well_indexD=well_indexD)
 
         return super().set_reservoir(reservoir)